feat: add simple fields validations on Controllers

+++ src/app/controllers/NaverController.js
@@ -1,16 +1,18 @@
 import Naver from '../models/Naver';
-import Projects from '../models/Project';
 
 class NaverController {
     async index(req, res) {
         const navers = await Naver.findAll();
+
         return res.json(navers);
     }
 
     async show(req, res) {
         const { naver_id } = req.params;
 
-        if (!naver_id) {
+        const findNaver = await Naver.findByPk(naver_id);
+
+        if (!findNaver) {
             return res.status(400).json({error: 'Naver not found!'});
         }
 
@@ -19,12 +21,14 @@ class NaverController {
         });
 
         return res.json(naver);
-    };
-
-    //Problemas ao adicionar um naver sem um projeto associado a ele. 
-    //Possiveis soluções: Enviar um array vazio no req.body, permitir nulo no model, validar esse dado...
+    };     
+    
     async store(req, res) {
         const { name, job_role, birth_date, admission_date, projects } = req.body;
+
+        if(!name || !job_role || !birth_date || !admission_date) {
+            return res.status(400).json({ error: 'Insert all required fields'});
+        }
 
         const naver = await Naver.create({
             name,
@@ -41,21 +45,43 @@ class NaverController {
     async update(req, res) {
         const { naver_id } = req.params;
 
+        const findNaver = await Naver.findByPk(naver_id);
+
+        if (!findNaver) {
+            return res.status(400).json({error: 'Naver not found!'});
+        }
+
         const [ naver ] = await Naver.update(req.body, {
             where: { id: naver_id },
         });
 
-        return res.json(naver);
+        if(naver >= 1) {
+            return res.json({msg: 'Naver changed with success!'});
+        } else {
+            return res.json({msg: 'Naver could not be changed.'});
+        }
+        
     }
 
     async delete(req, res) {
         const { naver_id } =  req.params;
 
+        const findNaver = await Naver.findByPk(naver_id);
+
+        if (!findNaver) {
+            return res.status(400).json({error: 'Naver not found!'});
+        }
+
         const deleted = await Naver.destroy({
             where: {id: naver_id}
         });
 
-        return res.json(deleted);
+        if(deleted >= 1) {
+            return res.json({msg: 'Naver deleted with success!'});
+        } else {
+            return res.json({msg: 'Naver could not be deleted.'});
+        }
+
     }
 }
